fixed pgenome_bin
git-svn-id: svn+ssh://combine-101/home/svn_bio/trunk@311 450597fd-3c47-0410-84d8-f85f2794411d

+++ probcalc_mp/probcalc_mp.c
@@ -851,8 +851,8 @@ inline int add_p_stats(mapping_t * fwd_map, mapping_t * rev_map,
 	if (discordant) {
 		pgenome = pgenome_fwd * pgenome_rev;
 	} else {
-		pgenome_bin = floor((ABS(double(dist) - gl_mean) 
-				* 1.0 / double(hist_distcutoff)) * double(HIST_BINS));
+		pgenome_bin = (int) floor((ABS(double(dist) - gl_mean) 
+				* 1.0 / hist_distcutoff) * HIST_BINS);
 	
 		if (pgenome_bin >= HIST_BINS)
 			pgenome_cumsum = 0;
